+++ packages/akashic-cli-serve/src/client/store/ContentEntity.ts
@@ -1,7 +1,8 @@
-import { observable } from "mobx";
+import { action, observable } from "mobx";
 import { SandboxConfig } from "../../common/types/SandboxConfig";
 import { ContentDesc } from "../../common/types/ContentDesc";
 import { ClientContentLocator } from "../common/ClientContentLocator";
+import * as ApiClient from "../api/ApiClient";
 
 export class ContentEntity {
 	readonly locator: ClientContentLocator;
@@ -16,5 +17,15 @@ export class ContentEntity {
 		Object.keys(args).forEach(key => {
 			this.argumentsTable[key] = JSON.stringify(args[key], null, 2);
 		});
+	}
+
+	@action
+	setSandboxConfig(config: SandboxConfig): void {
+		this.sandboxConfig = config;
+	}
+
+	async updateSandboxConfig(): Promise<void> {
+		const res = await ApiClient.getContent(this.locator.contentId);
+		this.setSandboxConfig(res.data.sandboxConfig || {});
 	}
 }